requirements.txt
asgiref==3.6.0
Django==4.1.6
django-browser-reload==1.6.0
django-environ==0.9.0
gunicorn==20.1.0
pytz==2022.7.1
sqlparse==0.4.3
